Create a metrics subpackage
